Market Data Service

+++ ValueScreener/ValueScreener/Controllers/StocksController.cs
@@ -3,19 +3,23 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Newtonsoft.Json;
 using ValueScreener.Data;
 using ValueScreener.Models;
 using ValueScreener.Models.Domain;
+using ValueScreener.Services.MarketData;
 
 namespace ValueScreener.Controllers
 {
     public class StocksController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly IMarketDataService _marketDataService;
 
-        public StocksController(ApplicationDbContext context)
+        public StocksController(ApplicationDbContext context, IMarketDataService marketDataService)
         {
             _context = context;
+            _marketDataService = marketDataService;
         }
 
         // GET: Stocks
@@ -91,15 +95,44 @@ namespace ValueScreener.Controllers
                 return NotFound();
             }
 
-            var stock = await _context.Stocks
+            var stock = await _context.Stocks.Include(s=>s.MarketData)
                 .SingleOrDefaultAsync(m => m.Id == id);
             if (stock == null)
             {
                 return NotFound();
             }
-
             return View(stock);
         }
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> RefreshMarketData(int?id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+            var stock = await _context.Stocks.Include(s => s.MarketData)
+                .SingleOrDefaultAsync(m => m.Id == id);
+            if (stock == null)
+            {
+                return NotFound();
+            }
+           
+            var marketData = await _marketDataService.GetMarketDataAsync(stock.Ticker);
+            if (marketData != null)
+            {
+                stock.MarketData = new MarketData
+                {
+                    LastPrice = marketData.LatestPrice,
+                    MarketCapitalization = marketData.MarketCap
+                };
+                if (marketData.LatestPrice.HasValue && marketData.MarketCap.HasValue)
+                    stock.MarketData.OutstandingShares = (long)Math.Floor(marketData.MarketCap.Value / marketData.LatestPrice.Value);
+                await _context.SaveChangesAsync();
+            }
+            return RedirectToAction(nameof(Details), new{id=id});
+        }
+
 
         // GET: Stocks/Create
         public IActionResult Create()
@@ -152,40 +185,38 @@ namespace ValueScreener.Controllers
         // POST: Stocks/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+        [HttpPost,ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Ticker,Name,Sector,Industry,Country,Currency,QuotationPlace")] Stock stock)
+        public async Task<IActionResult> EditPost(int ?id)
         {
-            if (id != stock.Id)
+            if (id == null)
             {
                 return NotFound();
             }
-
-            if (ModelState.IsValid)
+            var stockToUpdate = await _context.Stocks.SingleOrDefaultAsync(s => s.Id == id);
+            if (await TryUpdateModelAsync<Stock>(
+                stockToUpdate,
+                "",
+                s => s.Ticker, s => s.Name, s => s.Country, s=>s.Currency, s=>s.QuotationPlace, s=>s.Industry, s=>s.Sector))
             {
                 try
                 {
-                    _context.Update(stock);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateException /* ex */)
                 {
-                    if (!StockExists(stock.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    //Log the error (uncomment ex variable name and write a log.)
+                    ModelState.AddModelError("", "Unable to save changes. " +
+                                                 "Try again, and if the problem persists, " +
+                                                 "see your system administrator.");
                 }
-                return RedirectToAction(nameof(Index));
             }
-            return View(stock);
+            return View(stockToUpdate);
         }
 
         // GET: Stocks/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -193,10 +224,18 @@ namespace ValueScreener.Controllers
             }
 
             var stock = await _context.Stocks
+                .AsNoTracking()
                 .SingleOrDefaultAsync(m => m.Id == id);
             if (stock == null)
             {
                 return NotFound();
+            }
+
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                    "Delete failed. Try again, and if the problem persists " +
+                    "see your system administrator.";
             }
 
             return View(stock);
@@ -207,10 +246,25 @@ namespace ValueScreener.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var stock = await _context.Stocks.SingleOrDefaultAsync(m => m.Id == id);
-            _context.Stocks.Remove(stock);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            var stock = await _context.Stocks
+                .AsNoTracking()
+                .SingleOrDefaultAsync(m => m.Id == id);
+            if (stock == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            try
+            {
+                _context.Stocks.Remove(stock);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log.)
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
 
         private bool StockExists(int id)
